# Arec performance improvements