--- a/tests/Spec.hs
+++ b/tests/Spec.hs
@@ -37,7 +37,7 @@ main = hspec $ do
   describe "Fields may be accessed by overloaded labels" $ do
     it "Can get field X" $ rvalf #x d3 `shouldBe` 5
     it "Can get field Y" $ rvalf #y d3 `shouldBe` "Hi"
-  describe "ARec provides field accessors" $ do
+  describe "ARec" $ do
     it "Can get field Y" $ rvalf #y (toARec d3) `shouldBe` "Hi"
     it "Can set field X" $ rvalf #x (rputf #x 7 (toARec d3)) `shouldBe` 7
   describe "Converting between Rec and ARec" $ do