--- a/vinyl.cabal
+++ b/vinyl.cabal
@@ -24,6 +24,7 @@ library
   exposed-modules:     Data.Vinyl
                      , Data.Vinyl.ARec
                      , Data.Vinyl.ARec.Internal
+                     , Data.Vinyl.ARec.Internal.SmallArray
                      , Data.Vinyl.Class.Method
                      , Data.Vinyl.Core
                      , Data.Vinyl.CoRec
@@ -78,6 +79,9 @@ benchmark accessors
   hs-source-dirs:   benchmarks
   main-is:          AccessorsBench.hs
   build-depends:    base, criterion, tagged, vinyl, microlens
+  other-modules:    Bench.ARec
+                    Bench.SRec
+                    Bench.Rec
   ghc-options:      -O2
   default-language: Haskell2010