* Use `SmallArray#` instead of `Array` to avoid intermediate lists when constructing `ARec`s
  * Also saves a few words per record since we don't need Arrays' book-keeping
* Add ARec pseudo-constructors `&:`, `arnil` and `arec` to efficiently construct `ARec`s without going through `Rec`s
* Turn `ElField` into a newtype to improve accessor performance (both for `Rec` and `ARec`)
  * Also saves one word per field because we don't need the constructor
  * Also saves space for the `KnownSymbol` dictionary that wasn't used anywhere

See https://github.com/VinylRecords/Vinyl/issues/150#issuecomment-841847768 for benchmark comparisons

Example usage of the ARec pseudo-constructors: 

```haskell
user :: ARec ElField '[ "name"   ::: String
                      , "age"    ::: Int
                      , "active" ::: Bool]
user = arec (  #name   =: "Peter"
            &: #age    =: 4
            &: #active =: True
            &: arnil
            )
```


Related to #150 and #153