@Philonous 2021-05-19
Using the new `&:` and `arnil` constructors looks a lot like the old `toARec` method, compare with

```haskell
user = toArec ( #name   =: "Peter"
             :& #age    =: 4
             :& #active =: True
             :& RNil
             )
```

The difference is that it doesn't rely on inlining to get rid of the intermediate data structure, which might speed up compilation and work out better in -O0 situations, though I haven't tested it. I'm not sure if it's worth the additional code.