@acowley 2021-05-25
I've been thinking about this, and I'm leaning towards just keeping the old `toArec` variant. My main concern is the potential confusion between `&:` and `:&`. I think I did something similar to this (kind of cute variant of an operator) in another part of the API and quickly came to regret it as it was just too confusing (not in a deep way, of course, more of a USB plug kind of thing where you just have to try it both ways).