@Philonous 2021-05-25
I would agree, but I figured out that the same interface can be used for more things, e.g.

* A combinator that copies one arec into another, so you can easily add fields to an existing arec (either the beginning or the end)
* A combinator that copies a slice of an arec, so you can keep, say, the first 4 fields of an existing arec
* Updating an arec. The existing update combinator copies the entire array every time, with no way to batch updates.

But I would probably get away from the "cute" operators.