@Philonous 2021-05-26
I have removed `&:`, `arnil` and `arec` from the public interface. 

They still live in  `Data.Vinyl.ARec.Internal` since I used them to implement `toARec`. This should be much less of a problem, since Internal modules don't need to be stable, so they can be renamed or removed at will. I have also renamed `&:` to `arcons`. 

I have also added test cases for  `arecGetSubset` and `arecSetSubset` since I changed the code in a way that's not obviously semantically identical.