@acowley 2021-05-28
The comments you added are invaluable; I really appreciate that thoroughness!

The `Frames` test suite built against this new version of `vinyl` passed, but I'm hitting `Simplifier ticks exhausted` when building the [benchdemo](https://github.com/acowley/Frames/blob/master/benchmarks/BenchDemo.hs) executable.

The change in this PR that I'm most wary of is changing `ElField` to a `newtype`, which I'm guessing is related. I haven't really dug into it yet, and I'm interested to hear if you have some thoughts on what's going wrong.