@Philonous 2021-05-30
I can reproduce the compilation failure, and reverting 5e380747f5922160b486fef58ba1cc6993a3273c fixes it. 

How vexing, I would never have guess that a newtypes might lead to that problem.