@Philonous 2021-05-30
The problem might be some over-zealous inlining in Frames, though. 

I ran
```
sed -i -e 's/INLINE/INLINEABLE/g' **/*.hs
```
in the frames directory and it solved the problem