@Philonous 2021-05-30
In any case, the problem seems to be that turning the data into a newtype simplifies strictness analysis, giving the inliner more opportunities to fire.

I manually copied the TH-splice back into the source file, and commented out a bunch of column definitions, and the result is that reducing the number of column definitions drastically improves compilation times (almost exponentially). 

Btw, that's part of the reason that I am interested in ARecs, since they don't rely on unrolling, inlining so heavily.