@acowley 2021-05-30
That makes a lot of sense! I need to check that thE INLINABLE change doesn’t affect performance much.

Is there a way of determining steps we could take to ease the tradeoffs between compile- and run-times? I never worked out a way of profiling compile times, so I’m always flying a bit blind trying to help that. But I figured that — to some extent — inlining happens when something is used, so hopefully we’re not spending time optimizing something the user isn’t even using.