@Philonous 2021-05-30
OK, I bisected the `INLINE` pragmas in Frames until I found this single one which fixes the compilation issue for BenchDemo.hs:
https://github.com/acowley/Frames/blob/master/src/Frames/CSV.hs#L247

It doesn't seem to make a *huge* difference for run time:
before:
```
0.23user 0.00system 0:00.24elapsed 98%CPU
```
```
0.24user 0.01system 0:00.25elapsed 100%CPU
```