@acowley 2021-05-30
This is really interesting. I believe the thinking was that we wanted the whole chain of loading functions to be specialized to the right type. But perhaps it’s mostly okay if various layers are generic.

Many, perhaps most, uses of frames are some kind of fold. We want the inner part of the fold to in-line away into vector accesses of particular columns so there is no manifest `Rec`. Getting a better handle on where an opaque generic blocks that would be ideal.

For now, we should probably just remove that `INLINE` and keep moving forward. I’d really like to have a compilation benchmark harness so we can see the effects of changes on compile and run times. I’m not sure if there’s a way we could instruct cabal to help with that.