@Philonous 2021-05-31
Just to be a little more precise, I put the benchdemo `main` in a criterion harness and ran the different variants: 
[Gist](https://gist.github.com/Philonous/3c9b78961d5454f6356fbe410a139fed)

```
newtype / INLINEABLE: 194.3 ms   (192.0 ms .. 197.0 ms)
data / INLINEABLE 193.4 ms   (190.4 ms .. 198.1 ms)
data / INLINE 199.2 ms   (197.8 ms .. 199.8 ms)
data / all NOINLINE 199.1 ms   (197.3 ms .. 201.6 ms)
```

* newtype / data are the vinyl variants
* I tried both INLINE and  INLINEABLE on `readTableMaybeOpt`
* I replaced every occurence of `INLINE` with `NOINLINE` in the Frames code and ran the benchmark (last line)

It seems that the `INLINE` pragmas make no discernible difference in this particular case.