@Philonous 2021-05-31
Interestingly, comparing the core output from the `NOINLINE` version to the `INLINE` version of Benchdemo, I see this:

`data` + `NOINLINE`:
```
# grep -e ':&' -o BenchDemo.noinline.dump-simpl | wc -l
184
```
`data` + `INLINE`
```
grep -e ':&' -o BenchDemo.dump-simpl | wc -l
390
```

so it doesn't seem like the `INLINE` fused away the Rec in the first place, which would also explain why not inlining doesn't deteriorate performance