@acowley 2021-05-31
Okay, so one thing is to take out those `INLINE` pragmas in Frames. But of course then one has to wonder if we can do better. No difference between `newtype` and `data` is also a bit surprising given the `vinyl` benchmarks.

Is it worth making the `newtype` change now? My concern is that it can affect user code, and I’m not confident how it will benefit performance. It seems like the right thing to do if we can get away with it given what you found in `vinyl`, but I’d love to see it borne out in frames.