@Philonous 2021-06-01
Back of the envelope calculation: moving from data to newtype saved ~9ns for 1 row, 15 fields [1]
If we assume that everything scales linearly, then in the BenchDemo, which handles 36634 rows, 18 columns we'd expect to save
`9ns / 15 fields / 1row * 18 fields * 36634 rows = 395647ns = 0.4ms` 

That's below the run-to-run variance, so no wonder we don't see anything. 


[1]: https://github.com/VinylRecords/Vinyl/issues/150#issuecomment-841847768