@Philonous 2021-06-04
I've looked at `-dump-simpl-stats` for both the newtype and the data version. (Working on this project was the first time I had to deal with GHC debug output, so I'm not sure I'm reading it right)  Most of it looks like internally generated names that mean nothing to me, but one interesting difference was that `readRec` got inlined way more often in the newtype version. So I tried adding `{-# NOINLINE #-}` pragmas around it, which brought compilation time down from 13 to 9 seconds on my machine, but it also worsened the benchmark from 193ms to 216ms. 

It does however suggest that removing the right inline pragma(s) could make a significant difference in compilation times.

Regarding `SRec`: They are indeed as fast as normal Haskell records, but aren't they better compared to `UNPACK`ed Haskell records? As far as I remember these were faster still. But it's still impressive performance.